Second simulated data for the simple linear regression model

Second data used in example 4 of Salmerón, García and García (2024) (subsection 4.4) on the special case of the simple linear model.

Usage

data("SLM2")

Format

A data frame with 50 observations on the following 3 variables:

y2

Dependent variable simulated as y = 3 + 4*Z + u where u is normally distributed with a mean of 0 and a variance of 2.

cte

Intercept.

Z

Simulated from a normal distribution with a mean of 10 and a variance of 0.1.

References

Salmerón, R., García, C.B. and García, J. (2025). A redefined Variance Inflation Factor: overcoming the limitations of the Variance Inflation Factor. Computational Economics, 65, 337-363, doi: https://doi.org/10.1007/s10614-024-10575-8.

Examples

  head(SLM2, n=5)
#>         y2 cte         Z
#> 1 43.01204   1  9.978211
#> 2 40.04163   1  9.878235
#> 3 40.17086   1  9.924592
#> 4 40.79076   1 10.019123
#> 5 44.72774   1 10.104728
  y = SLM2[,1]
  x = SLM2[,2:3]
  multicollinearity(y, x)
#>        RVIFs         c0         c3 Scenario Affects
#> 1 187.800878 21.4798003 0.03277691      b.1     Yes
#> 2   1.879296  0.3687652 9.57724567      b.2      No