Soil characteristics data

Data used in Bondell and Reich's paper on soil characteristics used as predictors of forest diversity.

Usage

data("soil")

Format

A data frame with 20 observations on the following 16 variables.

BaseSat

% Base Saturation.

SumCation

Sum Cations (sums of cations like calcium, magnesium, potassium and sodium).

CECbuffer

CEC.

Ca

Calcium.

Mg

Magnesium.

K

Potassium.

Na

Sodium.

P

Phosphorus.

Cu

Copper.

Zn

Zinc.

Mn

Manganese.

HumicMatter

Humic Matter.

Density

Density.

pH

pH.

ExchAc

Exchangeable Acidity.

Diversity

Forest diversity (dependent variable).

Details

This dataset is originally used by Bondell and Reich (2008).

References

Bondell, H.D. and Reich. B.J. (2008). Simultaneous regression shrinkage, variable selection, and supervised clustering of predictors with OSCAR. Biometrics, 64 (1), 115–23, doi: https://doi.org/10.1111/j.1541-0420.2007.00843.x.

Examples

  head(soil, n=5)
#>   BaseSat SumCation CECbuffer     Ca     Mg      K    Na     P    Cu    Zn
#> 1    2.34    0.1576     0.614 0.0892 0.0328 0.0256 0.010 0.000 0.080 0.184
#> 2    1.64    0.0970     0.516 0.0454 0.0218 0.0198 0.010 0.000 0.064 0.112
#> 3    5.20    0.4520     0.828 0.3306 0.0758 0.0336 0.012 0.240 0.136 0.350
#> 4    4.10    0.3054     0.698 0.2118 0.0536 0.0260 0.014 0.030 0.126 0.364
#> 5    2.70    0.2476     0.858 0.1568 0.0444 0.0304 0.016 0.384 0.078 0.376
#>      Mn HumicMatter Density    pH ExchAc Diversity
#> 1 3.200      0.1220  0.0822 0.516  0.466 0.2765957
#> 2 2.734      0.0952  0.0850 0.512  0.430 0.2613982
#> 3 4.148      0.1822  0.0746 0.554  0.388 0.2553191
#> 4 3.728      0.1646  0.0756 0.546  0.408 0.2401216
#> 5 4.756      0.2472  0.0692 0.450  0.624 0.1884498
  y = soil[,16]
  x = soil[,-16] 
  x = cbind(rep(1, length(y)), x) # the design matrix has to have the intercept in the first column
  multicollinearity(y, x)
#> System is computationally singular. Modify the design matrix before running the code.
  multicollinearity(y, x[,-3]) # eliminating the problematic variable (SumCation)
#>           RVIFs           c0           c3 Scenario Affects
#> 1  4.407184e+02 6.150190e-03 1.480048e+00      b.1     Yes
#> 2  3.828858e+00 1.142356e-02 7.653413e+00      b.2      No
#> 3  1.093791e+05 1.254955e+02 7.236491e+04      b.1     Yes
#> 4  9.883235e+04 3.938383e+01 2.237445e+05      b.2      No
#> 5  1.767758e+05 1.101028e+03 3.609837e+05      b.2      No
#> 6  1.150029e+05 1.627349e+03 1.976176e+05      b.2      No
#> 7  4.627807e+04 5.960870e+02 2.033176e+06      b.2      No
#> 8  1.338591e+01 6.062571e-01 4.060382e+02      b.2      No
#> 9  3.113066e+02 4.089095e+01 5.246698e+05      b.2      No
#> 10 5.177176e+01 6.371216e+00 8.094828e+02      b.2      No
#> 11 1.905089e-01 3.907589e-02 9.787963e-01      b.2      No
#> 12 3.379360e+02 4.534540e+01 2.861964e+02      b.1     Yes
#> 13 4.761238e+04 8.453066e+01 3.828016e+08      b.2      No
#> 14 1.502903e+03 7.901580e+01 9.961215e+03      b.2      No
#> 15 1.066711e+05 2.369347e+02 4.802466e+07      b.2      No